fix(review): move classifyResolvedDecideError out of page.tsx — named export breaks Next.js page type-check (blocks Vercel prod deploy)
Vercel production deploy of main was failing — so the recent intake/review
fixes never went live (prod served a ~30h-old cached build). The Vercel build
log showed:

    Failed to type check.
    src/app/(workspace)/review/page.tsx
    Type error: Page "src/app/(workspace)/review/page.tsx" does not match the
    required types of a Next.js Page.
      "classifyResolvedDecideError" is not a valid Page export field.

Root cause: Next.js App Router allows a `page.tsx` to export ONLY the default
component plus a reserved set (metadata, generateMetadata, viewport, dynamic,
revalidate, ...). `review/page.tsx` also had a named `export function
classifyResolvedDecideError(...)` helper, which fails Next's page-type check.
The 18 required CI checks run vitest but NOT a production `next build`, so the
type-check only ran in the (non-required) Vercel deploy — the page passed every
gate while being unbuildable. PRE-EXISTING regression, not introduced by the
merge-chain.

Fix: move the pure helper into its own sibling module, decide-error.ts —
matching the existing route-local convention (destination.ts + destination.test.ts).
The decide-error.test.ts file already existed and tested it (importing from
"./page"); it now imports from "./decide-error". page.tsx imports the helper and
ends with ONLY `export default function ReviewPage()`. The function body +
docstring are unchanged.

Swept all mouth page.tsx/layout.tsx/template/error/loading/not-found files for
the same class of non-reserved named export: review/page.tsx was the ONLY
offender. (portal/.../messages/page.tsx does `export { default } from
"../chat/page"` — a default re-export, which is valid.) No second build error
of this class is lurking.

Verified locally (darwin):
- REPRO before fix: `npx next build --webpack` (cwd apps/mouth) -> exit 1 with
  the verbatim "classifyResolvedDecideError is not a valid Page export field".
- AFTER fix: `npx next build --webpack` -> exit 0 (compiled, type-check clean,
  230/230 static pages generated, full route table incl. /review printed).
- `npx vitest --run review/decide-error.test.ts` -> exit 0, 6/6 tests pass
  (classify logic still importable + tested from the new location).
