+++ apps/mouth/src/app/(workspace)/review/decide-error.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from "vitest";
 
-import { classifyResolvedDecideError } from "./page";
+import { classifyResolvedDecideError } from "./decide-error";
 
 describe("classifyResolvedDecideError", () => {
   it("treats an already-routed (filed) 409 as already_filed", () => {
